\name{MCE-package}
\alias{MCE-package}
\alias{MCE}
\docType{package}
\title{
Estimation of Monte Carlo Error
}
\description{
This package provides tools for estimating Monte Carlo error, and the appropriated number of replications in a Monte Carlo simulation study.
}
\details{
\tabular{ll}{
Package: \tab MCE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-12-05\cr
License: \tab BSD\cr
}

}
\author{
Elizabeth Koehler and Sebastien Haneuse

Maintainer: Elizabeth Koehler <e.koehler@vanderbilt.edu>

}
\references{

}

\keyword{ package }

\examples{
}
