% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{mesophotic}
\alias{mesophotic}
\title{Calculate the mesophotic community value for a given light level}
\usage{
mesophotic(light = seq(0.0065, 100, length = 500), a = 1.54, b = 8.92)
}
\arguments{
\item{light}{The percentage of surface irradiance (PAR) reaching a reef, positive and up to 100\%.}

\item{a}{The scale parameter of a Weibull distribution, defaults to (Tamir et al., 2019).
a describes the spread of the data, or ‘the peakyness’. Larger numbers create a flatter, wider, curve.
Varying a will simultaneously change the width of the preferred light environment of mesophotic taxa,
and the maximum similarity a reef patch is expected to achieve in comparison to the idealised mesophotic community.}

\item{b}{The shape parameter of a Weibull distribution, defaults to (Tamir et al., 2019). A value of 3
approximates a normal distribution. Values larger than 3 introduce a left-skew to the curve. This means
mesophotic taxa can be light-limited, while keeping mesophotic communities in low-light environments.
Varying b shifts the preferred light environment of mesophotic communities.}
}
\value{
A mesophotic community indicator value, as discussed in (Laverick et al., 2017). When a single light value is
provided, a single value is returned. You can pass a vector of light levels for a vector of depths.
}
\description{
This function captures the preferred light environment of the mesophotic coral community.
}
\details{
A Weibull distribution is used to represent the relationship between mesophotic coral communities
and light. We assume mesophotic community values (how similar a reef patch is to the idealised mesophotic
community) center on a preferred light value, between light limitation and light-induced stress.

This relationship has been verified in the Red Sea (Tamir et al., 2019), but certainly needs more research!
}
\examples{
# Specifying a single light value
# returns a single community value.

   mesophotic(light = 10)


# Specifying a vector of light values
# returns a vector of community values.

   mesophotic(light = seq(0.0065, 100, length = 500))


# By default the function uses the shallow community
# light relationship from (Tamir et al., 2019) for a
# range of light values.

   mesophotic()


# You can specify your own community light relationship
# by setting parameters explicitly

   mesophotic(a = 1.2, b = 8)


}
\references{
(Tamir et al., 2019)

(Laverick et al., 2017)
}
\seealso{
Other Community relationships: 
\code{\link{shallow}()}
}
\concept{Community relationships}
