% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{shallow}
\alias{shallow}
\title{Calculate the shallow community value for a given light level}
\usage{
shallow(light = seq(0.0065, 100, length = 500), Vm = 0.393, K = 13.5)
}
\arguments{
\item{light}{The percentage of surface irradiance (PAR) reaching a reef, positive and up to 100.}

\item{Vm}{The maximum similarity of a reef patch to the idealised shallow reef community, between 0 and 1.
Defaults to (Tamir et al., 2019).}

\item{K}{The light value where the community value = Vmax/2, between 0 and 100. Defaults to (Tamir et al., 2019).
This can be interpreted as the light level at which shallow reef communities become light limited.}
}
\value{
A shallow community indicator value, as discussed in (Laverick et al., 2017). When a single light value is
provided, a single value is returned. You can pass a vector of light levels for a vector of depths.
}
\description{
This function captures the preferred light environment of the "shallow" coral community.
}
\details{
A Michaelis-Menten equation is used to represent the relationship between shallow coral
communities and light. We assume shallow corals are light-limited beneath a light threshold, and that community
values (how similar a reef patch is to the idealised shallow community) increase asymptotically with increasing light.

This relationship has been verified in the Red Sea (Tamir et al., 2019), but certainly needs more research!
}
\examples{
# Specifying a single light value
# returns a single community value.

   shallow(light = 10)


# Specifying a vector of light values
# returns a vector of community values.

   shallow(light = seq(0.0065, 100, length = 500))


# By default the function uses the shallow community
# light relationship from (Tamir et al., 2019) for a
# range of light values.

   shallow()


# You can specify your own community light relationship
# by setting parameters explicitly

   shallow(Vm = 0.2, K = 11)


}
\references{
(Tamir et al., 2019)

(Laverick et al., 2017)
}
\seealso{
\code{\link{mesophotic}}
}
