\name{MCE-package}
\alias{MCE-package}
\alias{MCE}
\docType{package}
\title{
Estimation of Monte Carlo Error
}
\description{
This package provides tools for estimating Monte Carlo error, and the appropriated number of replications in a Monte Carlo \cr
 simulation study.
}
\details{
\tabular{ll}{
Package: \tab MCE\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2010-4-27\cr
License: \tab BSD\cr
}

}
\author{
Elizabeth Koehler and Sebastien Haneuse

Maintainer: Elizabeth Koehler <e.koehler@vanderbilt.edu>

}
\references{
  Koehler E, Brown E, Haneuse S:  On the assessment of Monte Carlo error in simulation-based statistical analyses. \cr
  \emph{The American Statistician}, \bold{63}(2): 155-162, 2009.
}

\keyword{ package }

