% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCE-package.R
\docType{package}
\name{MCE-package}
\alias{MCE}
\alias{MCE-package}
\title{MCE: Modelling Community Transitions Across Coral Reef Depth Gradients}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This package contains functions which predict the vertical distributions of coral reef communities. The package is based on the mechanistic model of (Laverick et al., 2020) which uses water clarity and community level light preferences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jack-h-laverick.github.io/MCE/}
  \item \url{https://github.com/Jack-H-Laverick/MCE}
  \item Report bugs at \url{https://github.com/Jack-H-Laverick/MCE/issues}
}

}
\author{
\strong{Maintainer}: Jack H Laverick \email{jacklaverick@ymail.com} (\href{https://orcid.org/0000-0001-8829-2084}{ORCID})

}
\keyword{internal}
